DEMOS-691: adjusted amendment names, flattened mock data
