+++ client/src/mock-data/index.ts
@@ -1,11 +1,6 @@
 import { demonstrationMocks } from "./demonstrationMocks";
 import { documentMocks } from "./documentMocks";
-import { eventMocks } from "./event/eventMocks";
+import { eventMocks } from "./eventMocks";
 import { userMocks } from "./userMocks";
 
-export const ALL_MOCKS = [
-  ...userMocks,
-  ...demonstrationMocks,
-  ...eventMocks,
-  ...documentMocks,
-];
+export const ALL_MOCKS = [...userMocks, ...demonstrationMocks, ...eventMocks, ...documentMocks];
